+++ src/main/java/de/allianz/notfallapp/dao/PositionDAO.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import de.allianz.notfallapp.entity.Position;
 
 @Transactional
-public interface PositionDAO extends CrudRepository<Position, Long>{
+public interface PositionDAO extends CrudRepository<Position, Integer>{
 
 }
