+++ src/main/java/de/allianz/notfallapp/dao/RoomDAO.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import de.allianz.notfallapp.entity.Room;
 
 @Transactional
-public interface RoomDAO extends CrudRepository<Room, Long> {
+public interface RoomDAO extends CrudRepository<Room, Integer> {
 
 }
